==> analysis/01_devo_curve_comparison.Rmd <==
---
title: "01_devo_curve_comparison.Rmd"
author: "anjie"
date: "`r Sys.Date()`"
output:
  pdf_document: default
  html_document: default
---

```{r}
library(tidyverse)
library(here)
library(ggthemes)
library(metafor)
library(ggrepel)


d <- read_csv(here("data/metalab_data_mini.csv")) %>% 
  filter(mean_age_months <= 36)

full_d <- read_csv(here("data/metalab_data_mini.csv")) 
source(here("analysis/helper/model_comparison_help.r"))
```





# Cacheing results   


## Estimates

```{r}
all_slope_estimates <- lapply(unique(d$ds_clean), 
       function(name){
         print(name)
         get_model_fit_df(d, name, NULL)
       }) %>% 
  bind_rows()

all_slope_estimates %>% saveRDS(here("cached_data/all_slope_estimates.Rds"))

```


## AIC comparison 

```{r}
age_models_df <- lapply(unique(d$ds_clean), 
       function(name){
         print(name)
         get_compare_IC_df(d, name, NULL)
       }) %>% 
  bind_rows()

saveRDS(age_models_df, here("cached_data/age_models_df.Rds"))
```


```{r}
full_age_models_df <- lapply(unique(full_d$ds_clean), 
       function(name){
         print(name)
         get_compare_IC_df(full_d, name, NULL)
       }) %>% 
  bind_rows()

```

## Model Predictions 

```{r}
min_age_df <- age_models_df %>% 
  filter(ic == "AICc") %>% 
  group_by(dataset) %>% 
  filter(REML == min(REML)) 

all_age_pred <- lapply(seq(1, nrow(min_age_df)), 
       function(x){
         get_age_model_prediction(min_age_df$dataset[[x]], min_age_df, d)
       }) %>% 
  bind_rows() 


all_age <- full_age_models_df %>% 
  filter(ic == "AICc") 

all_age_pred <- lapply(seq(1, nrow(all_age)), 
       function(x){
         get_all_age_model_prediction(all_age$dataset[[x]], all_age, all_age$model_spec[[x]], all_age$model_spec_clean[[x]], d)
       }) %>% 
  bind_rows() 




saveRDS(all_age_pred, here("cached_data/min_age_pred_df.RDS"))
saveRDS(all_age_pred, here("cached_data/full_age_pred_df.RDS"))

```




# Visualization 

```{r}
all_slope_estimates <- readRDS(here("cached_data/all_slope_estimates.Rds"))
age_models_df <- readRDS(here("cached_data/age_models_df.Rds"))
all_age_pred <- readRDS(here("cached_data/full_age_pred_df.RDS")) 
```

## prepping big plots 

```{r}

age_range_df <- d %>% 
  group_by(ds_clean) %>% 
  summarise(min_age = min(mean_age_months, na.rm = TRUE), 
            max_age = max(mean_age_months, na.rm = TRUE)) %>% 
  rename(ds_name = ds_clean)

categorized_age_pred_df <- all_age_pred %>% 
  mutate(
    model_type = case_when(
      !is.na(X.mean_age_months) ~ "linear", 
      !is.na(X.log.mean_age_months.) ~ "log", 
      !is.na(X.I.mean_age_months.2.) ~ "quad", 
      TRUE ~ "constant"
    )
  ) %>% 
  
  select(pred, ds_name, ci.ub, ci.lb, model_type) %>% 
  group_by(ds_name, model_type) %>% 
  mutate(mean_age_months = (row_number() / 2) - 0.5) %>% 
  left_join(age_range_df, by = c("ds_name")) %>% 
  rowwise() %>% 
  filter(if(ds_name != "Neonatal Imitation") mean_age_months >= min_age & mean_age_months < min(36, max_age) else mean_age_months >=0 & mean_age_months < min(36, max_age)) %>% 
  mutate(
    ds_type = case_when(
      ds_name %in% c( "Language discrimination and preference", 
                     "Mispronunciation sensitivity", 
                     "Natural speech preference", "Sound symbolism", "Statistical sound category learning", 
                     "Vowel discrimination (native)",
                     "Vowel discrimination (non-native)" ,"Word Segmentation (combined)",
                     "Statistical word segmentation", "Audio-Visual Congruence", 
                     "Language discrimination and preference (discrimination)", 
                     "Language discrimination and preference (preference)") ~ "Sounds", 
      
      ds_name %in% c("Familiar word recognition", 
                     "Cross-situational word learning", 
                      "Label advantage in concept learning",
                     "Mutual exclusivity",
                     "Online word recognition",
                     "Syntactic bootstrapping",
                     "Switch task") ~ "Words",
      
      ds_name %in% c("Gaze following (combined)",
                     "Neonatal Imitation",
                     "Prosocial agents",
                     #"Symbolic play", 
                     "Infant directed speech preference") ~ "Communication", 
      
      ds_name %in% c("Abstract rule learning",  
                     "Categorization bias", 
                     "Simple arithmetic competences") ~ "Cognitive"
     )
  ) %>% 
  mutate(pred = case_when(
    ds_name == "Abstract rule learning" ~ -pred, 
    ds_name == "Switch task" ~ -pred,
    TRUE ~ pred
  ), 
  ci.ub = 
    case_when(
    ds_name == "Abstract rule learning" ~ -ci.lb, 
    ds_name == "Switch task" ~ -ci.lb,
    TRUE ~ ci.ub
  ),
   ci.lb = 
    case_when(
    ds_name == "Abstract rule learning" ~ -ci.ub, 
    ds_name == "Switch task" ~ -ci.ub,
    TRUE ~ ci.lb
  )
  
  ) %>% 
  select(ds_name, ds_type, mean_age_months, pred, ci.ub, ci.lb, model_type) %>% 
  ungroup()

significant_res = all_slope_estimates %>% 
  filter(term != "intercept") %>% 
  filter(p.value < .05) %>% 
  pull(dataset)

categorized_d <- d %>% 
  mutate(
    ds_type = case_when(
      ds_clean %in% c( "Language discrimination and preference", 
                     "Mispronunciation sensitivity", 
                     "Natural speech preference", "Sound symbolism", "Statistical sound category learning", 
                     "Vowel discrimination (native)",
                     "Vowel discrimination (non-native)" ,"Word Segmentation (combined)",
                     "Statistical word segmentation", "Audio-Visual Congruence", 
                     "Language discrimination and preference (discrimination)", 
                     "Language discrimination and preference (preference)") ~ "Sounds", 
      
      ds_clean %in% c("Familiar word recognition", 
                     "Cross-situational word learning", 
                      "Label advantage in concept learning",
                     "Mutual exclusivity",
                     "Online word recognition",
                     "Syntactic bootstrapping",
                     "Switch task") ~ "Words",
      
      ds_clean %in% c("Gaze following (combined)",
                     "Neonatal Imitation",
                     "Prosocial agents",
                     #"Symbolic play", 
                     "Infant directed speech preference") ~ "Communication", 
      
      ds_clean %in% c("Abstract rule learning",  
                     "Categorization bias", 
                     "Simple arithmetic competences") ~ "Cognitive"
     )
  ) %>% 
  mutate( 
    d_calc = case_when(
      ds_clean == "Abstract rule learning" ~ -d_calc, 
    ds_clean == "Switch task" ~ -d_calc,
    TRUE ~ d_calc))

categorized_d <- categorized_d %>% 
  rename(ds_name = ds_clean) %>% 
  mutate(
     ds_name = case_when(
      ds_name == "Cross-situational word learning" ~ "X-situation learning", 
      ds_name == "Label advantage in concept learning" ~ "Label advantage",
      ds_name == "Language discrimination and preference (discrimination)" ~ "Language discrimination",       ds_name == "Audio-Visual Congruence" ~ "Audio-visual congruen.", 

      ds_name == "Language discrimination and preference (preference)" ~ "Language preference",
      ds_name == "Cross-situational word learning"  ~ "X-situation learning",
      ds_name == "Gaze following (combined)" ~ "Gaze following", 
      ds_name == "Label advantage in concept learning" ~ "Label advantage", 
      ds_name == "Statistical sound category learning" ~ "Stat. category learning",
      ds_name == "Statistical word segmentation" ~ "Stat. word seg.",
      ds_name == "Word Segmentation (combined)" ~ "Word segmentation", 
      ds_name == "Infant directed speech preference" ~ "IDS pref.",
      TRUE ~ ds_name
  )) #%>% 
  #mutate(
  #  ds_name = case_when(
  #    ds_name %in% significant_res ~ paste0(ds_name, "*"), 
  #    TRUE ~ ds_name
  #  )
  #)
```

## making big plot 

```{r, fig.width=9, fig.height=10}
categorized_age_pred_df %>% 
  # this is to handle the tricky part with neonatal imitation
  filter(pred != -Inf) %>% 
  mutate(group_var = paste0(ds_name, model_type)) %>% 
  mutate(
     ds_name = case_when(
       ds_name == "Cross-situational word learning" ~ "X-situation learning", 
      ds_name == "Audio-Visual Congruence" ~ "Audio-visual congruen.", 
      ds_name == "Label advantage in concept learning" ~ "Label advantage",
      ds_name == "Language discrimination and preference (discrimination)" ~ "Language discrimination", 
      ds_name == "Language discrimination and preference (preference)" ~ "Language preference",
      ds_name == "Cross-situational word learning"  ~ "X-situation learning",
      ds_name == "Gaze following (combined)" ~ "Gaze following", 
      ds_name == "Label advantage in concept learning" ~ "Label advantage", 
      ds_name == "Statistical sound category learning" ~ "Stat. category learning",
      ds_name == "Statistical word segmentation" ~ "Stat. word seg.",
      ds_name == "Word Segmentation (combined)" ~ "Word segmentation", 
      ds_name == "Infant directed speech preference" ~ "IDS pref.",
      TRUE ~ ds_name
  )) %>% 
  mutate(
    model_type = case_when(
      model_type == "constant" ~ "Constant", 
      model_type == "linear" ~ "Linear", 
      model_type == "quad" ~ "Quadratic", 
      model_type == "log" ~ "Logarithmic"
    )
  ) %>% 

  ggplot(aes(x = mean_age_months,  y = pred)) + 
  geom_hline(yintercept = 0, linetype ="dotted", alpha = .5) + 
  geom_point(data = categorized_d , 
            aes(x = mean_age_months, y = d_calc),size = .5, alpha = .5) + 
  geom_line(alpha = 1,  aes(group = group_var,color = model_type)) + 
  
 geom_ribbon(
              aes(x = mean_age_months, group = group_var,y = pred, ymin = ci.lb, ymax = ci.ub, fill = model_type), alpha = 0.2, colour = NA)+ 

    theme_few()+
     guides(color=guide_legend(title="Model Type")) + 
guides(fill = "none") + 
xlab("Mean age (months)") + 
  ylab("Predicted ES") + 
  
   theme(legend.position = "top", 
         strip.text = element_text(size = 11, margin = margin())) + 
  
  facet_wrap(~ds_name, scales = "free_x",
             labeller = label_wrap_gen()) 
```


## Slope historgram

```{r}
all_slope_estimates %>% 
  filter(term != "intercept") %>% 
  ggplot(aes(x = estimate)) + 
  geom_histogram(bins = 40) + 
  #geom_density()+
  theme_few()
```

```{r}
all_slope_estimates %>% 
  filter(term != "intercept") %>% 
  mutate(significant = p.value < .05) %>% 
  ggplot(aes(x = reorder(dataset, estimate), y = estimate, ymin = conf.low, ymax = conf.high, color = significant)) + 
  scale_color_manual(values = c("black", "red")) + 
  geom_hline(yintercept = 0, linetype = "dashed") + 
  geom_pointrange() + 
  coord_flip() + 
  theme_few() + 
  xlab("")
```



## AICc Table 

```{r}
age_models_df %>% 
  filter(ic == "AICc") %>% 
  rename(Dataset = dataset) %>% 
  select(REML, model_spec_clean, Dataset) %>% 
  pivot_wider(names_from = model_spec_clean, 
              values_from = REML) %>% 
  kableExtra::kable(digits = 2)
```


```{r}
age_models_df %>% 
  filter(ic == "AICc") %>% 
  rename(Dataset = dataset) %>% 
  select(REML, model_spec_clean, Dataset) %>% 
  pivot_wider(names_from = model_spec_clean, 
              values_from = REML) 
```


## AICc Table Visualization

```{r}

age_df_wide <- age_models_df %>% 
  filter(ic == "AICc") %>% 
  rename(Dataset = dataset) %>% 
  select(REML, model_spec_clean, Dataset) %>% 
  pivot_wider(names_from = model_spec_clean, 
              values_from = REML)

age_df_wide$min <- apply(age_df_wide[c('Linear','Log', 'Quadratic', 'Const')], 1, min)

# get the delta value 
age_df_wide <- age_df_wide %>% 
  mutate(d_linear = Linear - min, 
         d_log = Log - min, 
         d_quad = Quadratic - min, 
         d_const= Const - min) %>% 
  select(Dataset, d_linear, d_log, d_quad, d_const) %>% 
  rename(Linear = d_linear, 
         Log = d_log, 
         Quadratic = d_quad, 
         Const = d_const) %>% 
   mutate(across(where(is.numeric), round, 2)) 


age_df_wide %>% 
  pivot_longer(cols = c("Linear", "Log", "Quadratic", "Const")) %>% 
  ggplot(aes(x = reorder(Dataset, value), y = value, color = name)) +
  theme_few() + 
  geom_hline(yintercept = 4, linetype = 2, color = "gray") + 
  geom_hline(yintercept = 0, linetype = 3) + 
  geom_point(position = position_dodge(width = .2), 
             #color = "white"
             ) + 

  coord_flip()+ 
  xlab("") + 
  labs(color = "Functional form") + 
  guides(colour = guide_legend(nrow = 2))  + 
  theme(legend.position = "top")
```


```{r}
full_age_df_wide <- full_age_models_df %>% 
  filter(ic == "AICc") %>% 
  rename(Dataset = dataset) %>% 
  select(REML, model_spec_clean, Dataset) %>% 
  pivot_wider(names_from = model_spec_clean, 
              values_from = REML)

full_age_df_wide$min <- apply(full_age_df_wide[c('Linear','Log', 'Quadratic', 'Const')], 1, min)

# get the delta value 
full_age_df_wide <- full_age_df_wide %>% 
  mutate(d_linear = Linear - min, 
         d_log = Log - min, 
         d_quad = Quadratic - min, 
         d_const= Const - min) %>% 
  select(Dataset, d_linear, d_log, d_quad, d_const) %>% 
  rename(Linear = d_linear, 
         Log = d_log, 
         Quadratic = d_quad, 
         Const = d_const) %>% 
   mutate(across(where(is.numeric), round, 2)) 


full_age_df_wide %>% 
  pivot_longer(cols = c("Linear", "Log", "Quadratic", "Const")) %>% 
  ggplot(aes(x = reorder(Dataset, value), y = value, color = name)) +
  theme_few() + 
  geom_hline(yintercept = 4, linetype = 2, color = "gray") + 
  geom_hline(yintercept = 0, linetype = 3) + 
  geom_point(position = position_dodge(width = .2), 
             #color = "white"
             ) + 

  coord_flip()+ 
  xlab("") + 
  labs(color = "Functional form") + 
  guides(colour = guide_legend(nrow = 2))  + 
  theme(legend.position = "top")
```

```{r}
full_age_df_wide %>% 
  pivot_longer(cols = c("Linear", "Log", "Quadratic", "Const")) %>% 
  mutate(type = "Full age range") %>% 
  bind_rows(
    age_df_wide %>% 
    pivot_longer(cols = c("Linear", "Log", "Quadratic", "Const")) %>% 
    mutate(type = "36MO") 
  ) %>% 
  ggplot(aes(x = reorder(Dataset, value), y = value, color = name, shape = type, group = type)) +
  theme_few() + 
  geom_hline(yintercept = 4, linetype = 2, color = "gray") + 
  geom_hline(yintercept = 0, linetype = 3) + 
  geom_point(position = position_dodge(width = .8), 
             #color = "white"
             ) + 

  coord_flip()+ 
  xlab("") + 
  labs(color = "Functional form") + 
  guides(colour = guide_legend(nrow = 2))  


```



```{r}
age_models_df %>% 
  filter(ic == "AICc") %>% 
  rename(Dataset = dataset) %>% 
  select(REML, model_spec_clean, Dataset) %>% 
  pivot_wider(names_from = model_spec_clean, 
              values_from = REML) %>% 
  mutate(
    linear = Linear - Const,
    log = Log - Const, 
    quad = Quadratic - Const
  ) %>% 
  select(Dataset, log, quad, linear) %>% 
  filter(log > 4 | quad > 4 | linear > 4)
```


4 different results: 

Categorization bias: fixed random effect structure (add same_infant)
Statistical sound category learning (habituation): changed random effect structure (add same_infant)
Infant directed speech preference: added unique_row 
Mispronunciation sensitivity: fixed two missing infants age 



```{r}
age_diff_df <- d %>% 
  select(ds_clean, mean_age_months) %>% 
  group_by(ds_clean) %>% 
  arrange(mean_age_months,.by_group = TRUE) %>% 
  mutate(age_diff = mean_age_months - lag(mean_age_months)) %>% 
  filter(!is.na(age_diff)) %>% 
  group_by(ds_clean) %>% 
  tidyboot::tidyboot_mean(age_diff)
```
```{r}
age_diff_df %>% 
  ggplot(aes(x = reorder(ds_clean, -mean), y = mean, ymin = ci_lower, ymax = ci_upper)) + 
  geom_pointrange() + 
  coord_flip() + 
  theme_few()
```


## Devo Curve 



```{r}
categorized_d <- d %>% 
  mutate(
    ds_type = case_when(
      ds_clean %in% c( "Language discrimination and preference", 
                     "Mispronunciation sensitivity", 
                     "Natural speech preference", "Sound symbolism", "Statistical sound category learning", 
                     "Vowel discrimination (native)",
                     "Vowel discrimination (non-native)" ,"Word Segmentation (combined)",
                     "Statistical word segmentation", "Audio-Visual Congruence", 
                     "Language discrimination and preference (discrimination)", 
                     "Language discrimination and preference (preference)") ~ "Sounds", 
      
      ds_clean %in% c("Familiar word recognition", 
                     "Cross-situational word learning", 
                      "Label advantage in concept learning",
                     "Mutual exclusivity",
                     "Online word recognition",
                     "Syntactic bootstrapping",
                     "Switch task") ~ "Words",
      
      ds_clean %in% c("Gaze following (combined)",
                     "Neonatal Imitation",
                     "Prosocial agents",
                     #"Symbolic play", 
                     "Infant directed speech preference") ~ "Communication", 
      
      ds_clean %in% c("Abstract rule learning",  
                     "Categorization bias", 
                     "Simple arithmetic competences") ~ "Cognitive"
     )
  ) %>% 
  mutate( 
    d_calc = case_when(
      ds_clean == "Abstract rule learning" ~ -d_calc, 
    ds_clean == "Switch task" ~ -d_calc,
    TRUE ~ d_calc))
  
```

```{r}


age_range_df <- d %>% 
  group_by(ds_clean) %>% 
  summarise(min_age = min(mean_age_months, na.rm = TRUE), 
            max_age = max(mean_age_months, na.rm = TRUE)) %>% 
  rename(ds_name = ds_clean)

categorized_age_pred_df <- all_age_pred %>% 
  mutate(
    model_type = case_when(
      !is.na(X.mean_age_months) ~ "linear", 
      !is.na(X.log.mean_age_months.) ~ "log", 
      !is.na(X.I.mean_age_months.2.) ~ "quad", 
      TRUE ~ "constant"
    )
  ) %>% 
  
  select(pred, ds_name, ci.ub, ci.lb, model_type) %>% 
  group_by(ds_name, model_type) %>% 
  mutate(mean_age_months = (row_number() / 2) - 0.5) %>% 
  left_join(age_range_df, by = c("ds_name")) %>% 
  rowwise() %>% 
  filter(if(ds_name != "Neonatal Imitation") mean_age_months >= min_age & mean_age_months < min(36, max_age) else mean_age_months >=0 & mean_age_months < min(36, max_age)) %>% 
  mutate(
    ds_type = case_when(
      ds_name %in% c( "Language discrimination and preference", 
                     "Mispronunciation sensitivity", 
                     "Natural speech preference", "Sound symbolism", "Statistical sound category learning", 
                     "Vowel discrimination (native)",
                     "Vowel discrimination (non-native)" ,"Word Segmentation (combined)",
                     "Statistical word segmentation", "Audio-Visual Congruence", 
                     "Language discrimination and preference (discrimination)", 
                     "Language discrimination and preference (preference)") ~ "Sounds", 
      
      ds_name %in% c("Familiar word recognition", 
                     "Cross-situational word learning", 
                      "Label advantage in concept learning",
                     "Mutual exclusivity",
                     "Online word recognition",
                     "Syntactic bootstrapping",
                     "Switch task") ~ "Words",
      
      ds_name %in% c("Gaze following (combined)",
                     "Neonatal Imitation",
                     "Prosocial agents",
                     #"Symbolic play", 
                     "Infant directed speech preference") ~ "Communication", 
      
      ds_name %in% c("Abstract rule learning",  
                     "Categorization bias", 
                     "Simple arithmetic competences") ~ "Cognitive"
     )
  ) %>% 
  mutate(pred = case_when(
    ds_name == "Abstract rule learning" ~ -pred, 
    ds_name == "Switch task" ~ -pred,
    TRUE ~ pred
  ), 
  ci.ub = 
    case_when(
    ds_name == "Abstract rule learning" ~ -ci.lb, 
    ds_name == "Switch task" ~ -ci.lb,
    TRUE ~ ci.ub
  ),
   ci.lb = 
    case_when(
    ds_name == "Abstract rule learning" ~ -ci.ub, 
    ds_name == "Switch task" ~ -ci.ub,
    TRUE ~ ci.lb
  )
  
  ) %>% 
  select(ds_name, ds_type, mean_age_months, pred, ci.ub, ci.lb, model_type) %>% 
  ungroup()

```


## current try out w/o label 
```{r}


label_df <- categorized_age_pred_df %>% 
  group_by(ds_name) %>% 
  filter(mean_age_months == max(mean_age_months), model_type == "linear") %>%
  distinct(ds_name, mean_age_months, pred, ds_type) 

#categorized_age_pred_df %>% filter(ds_type == "Words") %>% distinct(ds_name)
  

label_df$x = label_df$mean_age_months 
label_df$y = label_df$pred 

label_df <- label_df %>% 
  mutate(
    x = case_when(
      ds_name == "Categorization bias" ~ x + 1, 
      ds_name == "Neonatal Imitation" ~ x + 1,
      ds_name == "Prosocial agents" ~ x + 1,
      ds_name == "Mutual exclusivity" ~ x - 3, 
      ds_name == "Cross-situational word learning" ~ x -3,
      ds_name == "Label advantage in concept learning" ~ x - 10, 
      ds_name == "Syntactic bootstrapping" ~ x-1,
      ds_name == "Vowel discrimination (native)" ~ x + 1.5,
      ds_name == "Statistical word segmentation"  ~ x + 1,
      ds_name == "Audio-Visual Congruence" ~ x - 7, 
       ds_name == "Language discrimination and preference (preference)" ~ x + 5,
      ds_name == "Natural speech preference" ~ x + 6,
       ds_name == "Statistical sound category learning" ~ x +6,
      ds_name == "Vowel discrimination (native)" ~ x - .8,
      ds_name == "Vowel discrimination (non-native)" ~ x - 10,
      ds_name == "Sound symbolism" ~ x -.4,
    
     
      TRUE ~ x
    ), 
    
    y = case_when(
      ds_name == "Categorization bias" ~ y + 2, 
      ds_name == "Simple arithmetic competences" ~ y + .5,
      ds_name == "Abstract rule learning" ~ y - .5,
      ds_name == "Gaze following (combined)" ~ y + .5,
      ds_name == "Prosocial agents" ~ y + .5,
      ds_name == "Infant directed speech preference" ~ y - .5,
      ds_name == "Cross-situational word learning" ~ y + .5,
      ds_name == "Familiar word recognition" ~ y + .5,
      ds_name == "Label advantage in concept learning" ~ y - .8, 
      ds_name == "Syntactic bootstrapping" ~ y-.5,
       ds_name == "Switch task" ~ y + .3,
      ds_name == "Mispronunciation sensitivity" ~ y + .5,
      ds_name == "Vowel discrimination (native)" ~ y - .45,
       ds_name == "Statistical word segmentation"  ~ y - .5,
      ds_name == "Statistical sound category learning" ~ y+.1,
      ds_name == "Natural speech preference" ~ y + .8,
      ds_name == "Audio-Visual Congruence" ~ y + .8,
      ds_name == "Word Segmentation (combined)" ~ y - .4, 
      ds_name == "Language discrimination and preference (preference)" ~ y + .8,
       ds_name == "Vowel discrimination (non-native)" ~ y + .8,
       ds_name == "Sound symbolism" ~ y + .3,
      TRUE ~ y
    ), 
    
    label = case_when(
      ds_name == "Cross-situational word learning" ~ "X-situation learning", 
      ds_name == "Label advantage in concept learning" ~ "Label advantage",
      ds_name == "Language discrimination and preference (discrimination)" ~ "Language discrimination", 
      ds_name == "Language discrimination and preference (preference)" ~ "Language preference",
      ds_name == "Cross-situational word learning"  ~ "X-situation learning",
      ds_name == "Gaze following (combined)" ~ "Gaze following", 
      ds_name == "Label advantage in concept learning" ~ "Label advantage", 
      ds_name == "Statistical sound category learning" ~ "Stats. category learning",
      ds_name == "Statistical word segmentation" ~ "Stats. word seg.",
      ds_name == "Word Segmentation (combined)" ~ "Word segmentation", 
      ds_name == "Infant directed speech preference" ~ "IDS pref.",
      TRUE ~ ds_name
    )
  ) #%>% 
  #filter(label == "Vowel discrimination (native)")
      
   




# 
# 
#   c(
#     13 + 4,#Abstract rule learning
#     0 + 8, # audio visual
#     35.5 + 5 , # Categorization bias
#     35.5 + 3, # "Cross-situational word learning"
#     15 + 3,#Familiar word recognition
#     23.5 + 4, #"Gaze following (combined)",,
#     15 - 3, # Label advantage
#     11.5 + 8, # "Language discrimination and preference"
#     35.5 + 4,#"Mispronunciation sensitivity"
#     35.5 + 4, #Mutual exclusivity , 
#     12.5 , #"Natural speech preference"
#     0 + 4, #neonatal imitation 
#     30 + 2, #"Online word recognition"
#     31.5 + 8, #Prosocial Agents
#     9 + 4, #"Simple arithmetic competences"
#     35.5 + 8, #Sound symbolism, 
#     11 + 5, #"Statistical sound category learning (habituation)"
#     18.5, #"Statistical word segmentation"
#     24.5+ 3, #Switch task,
#     35.5 + 5, # "Syntactic bootstrapping"
#     17.5 - 5, #"Vowel discrimination (non-native)"
#     29.5 + 11, #"Vowel discrimination (native)"
#     24.5 + 14, #"Word Segmentation (combined)"
#     18.5# "Infant directed speech preference"
# )
# label_df$y = c(
#   0.05 - 0.5,#Abstract rule learning
#   0 + 1.5, # audio visual
#   0.275 + .4, # Categorization bias
#   0.545 + .3, # "Cross-situational word learning"
#   0.805 + 0.3,#Familiar word recognition
#   2.99 + 0.3, #"Gaze following (combined)"
#   0.05004268 - 0.5, # label advantage
#   0.265 - 1, # "Language discrimination and preference"
#   1.62 + 0.5,#"Mispronunciation sensitivity"
#   1.44 + 0.3,#Mutual exclusivity
#   0.811 + 1.3, #"Natural speech preference"
#   0 + 1.5, # neonatal imitation 
#   2.964331 + 0.5, #"Online word recognition"
#   0.405, #Prosocial agents
#   0.249 + 0.6,#"Simple arithmetic competences"
#   0.383, #Sound symbolism, 
#   0.557 + 0.5, #"Statistical sound category learning (habituation)"
#   -0.07808538 - 0.2, #"Statistical word segmentation"
#   0.156 - 0.3,#Switch task
#   0.103 - 0.5, # "Syntactic bootstrapping"
#   0.5901183 + 0.2, #"Vowel discrimination (native)"
#   0.6457766 + 1, #"Vowel discrimination (non-native)"
#   0.2 + 1.2, #"Word Segmentation (combined)"
#  0.467 - 0.6# "Infant directed speech preference", 
#  #0.627 + 0.5, #Symbolic play
# )


categorized_age_pred_df %>% 
  mutate(group_var = paste0(ds_name, model_type)) %>% 
  group_by(
    ds_name
  ) %>% 
  mutate(
    label =  if_else(mean_age_months == max(mean_age_months) & model_type == "linear", as.character(ds_name), NA_character_)
  ) %>% 
  ungroup() %>% 
 #filter(label == "Switch task") %>% 
  ggplot(aes(x = mean_age_months,  y = pred,  color = ds_name)) + 
  geom_hline(yintercept = 0, linetype ="dashed", alpha = .5) + 
  geom_smooth(data = categorized_d, #%>% filter(ds_clean == "Vowel discrimination (native)"), 
              aes(x = mean_age_months, y = d_calc, group = ds_clean, color = ds_clean), 
              method = "loess", span = 2,  se = FALSE)+ 
  geom_line(alpha = .5, linetype = "dashed", aes(group = group_var)) + 
   

 geom_ribbon(
              aes(x = mean_age_months, group = group_var,y = pred, ymin = ci.lb, ymax = ci.ub, fill = ds_name), alpha = 0.05, colour = NA)+ 

    theme_few()+
    geom_label(data = label_df, aes(x = x, y = y, label = label), 
                   size = 2) + 

xlab("Mean age (months)") + 
  ylab("Predicted ES") + 
  
   theme(legend.position = "none") + 
  facet_wrap(~ds_type, scales = "free_y")  

```


```{r, fig.width=8,fig.height = 8}

significant_res = all_slope_estimates %>% 
  filter(term != "intercept") %>% 
  filter(p.value < .05) %>% 
  pull(dataset)

categorized_d <- d %>% 
  mutate(
    ds_type = case_when(
      ds_clean %in% c( "Language discrimination and preference", 
                     "Mispronunciation sensitivity", 
                     "Natural speech preference", "Sound symbolism", "Statistical sound category learning", 
                     "Vowel discrimination (native)",
                     "Vowel discrimination (non-native)" ,"Word Segmentation (combined)",
                     "Statistical word segmentation", "Audio-Visual Congruence", 
                     "Language discrimination and preference (discrimination)", 
                     "Language discrimination and preference (preference)") ~ "Sounds", 
      
      ds_clean %in% c("Familiar word recognition", 
                     "Cross-situational word learning", 
                      "Label advantage in concept learning",
                     "Mutual exclusivity",
                     "Online word recognition",
                     "Syntactic bootstrapping",
                     "Switch task") ~ "Words",
      
      ds_clean %in% c("Gaze following (combined)",
                     "Neonatal Imitation",
                     "Prosocial agents",
                     #"Symbolic play", 
                     "Infant directed speech preference") ~ "Communication", 
      
      ds_clean %in% c("Abstract rule learning",  
                     "Categorization bias", 
                     "Simple arithmetic competences") ~ "Cognitive"
     )
  ) %>% 
  mutate( 
    d_calc = case_when(
      ds_clean == "Abstract rule learning" ~ -d_calc, 
    ds_clean == "Switch task" ~ -d_calc,
    TRUE ~ d_calc))

categorized_d <- categorized_d %>% 
  rename(ds_name = ds_clean) %>% 
  mutate(
     ds_name = case_when(
      ds_name == "Cross-situational word learning" ~ "X-situation learning", 
      ds_name == "Label advantage in concept learning" ~ "Label advantage",
      ds_name == "Language discrimination and preference (discrimination)" ~ "Language discrimination", 
      ds_name == "Language discrimination and preference (preference)" ~ "Language preference",
      ds_name == "Cross-situational word learning"  ~ "X-situation learning",
      ds_name == "Gaze following (combined)" ~ "Gaze following", 
      ds_name == "Label advantage in concept learning" ~ "Label advantage", 
      ds_name == "Statistical sound category learning" ~ "Stats. category learning",
      ds_name == "Statistical word segmentation" ~ "Stats. word seg.",
      ds_name == "Word Segmentation (combined)" ~ "Word segmentation", 
      ds_name == "Infant directed speech preference" ~ "IDS pref.",
      TRUE ~ ds_name
  )) %>% 
  mutate(
    ds_name = case_when(
      ds_name %in% significant_res ~ paste0(ds_name, "*"), 
      TRUE ~ ds_name
    )
  )

age_range_df <- d %>% 
  group_by(ds_clean) %>% 
  summarise(min_age = min(mean_age_months, na.rm = TRUE), 
            max_age = max(mean_age_months, na.rm = TRUE)) %>% 
  rename(ds_name = ds_clean)


sig_ds <- c("Mutual exclusivity", "Audio-Visual Congruence", "Mispronunciation sensitivity",
           "Statistical sound category learning (habituation)", "Simple arithmetic competences")

categorized_age_pred_df <- all_age_pred %>% 
  select(pred, ds_name, ci.ub, ci.lb) %>% 
  group_by(ds_name) %>% 
  mutate(mean_age_months = (row_number() + 1) / 2) %>% 
  left_join(age_range_df, by = c("ds_name")) %>% 
  filter(mean_age_months > min_age, 
         mean_age_months < min(36, max_age)) %>% 
  mutate(
    ds_type = case_when(
      ds_name %in% c( "Language discrimination and preference", 
                     "Mispronunciation sensitivity", 
                     "Natural speech preference", "Sound symbolism", "Statistical sound category learning (habituation)", 
                     "Vowel discrimination (native)",
                     "Vowel discrimination (non-native)" ,"Word Segmentation (combined)",
                     "Statistical word segmentation", "Audio-Visual Congruence") ~ "Sounds", 
      
      ds_name %in% c("Familiar word recognition", 
                     "Cross-situational word learning", 
                      "Label advantage in concept learning",
                     "Mutual exclusivity",
                     "Online word recognition",
                     "Syntactic bootstrapping",
                     "Switch task") ~ "Words",
      
      ds_name %in% c("Gaze following (combined)",
                     "Neonatal Imitation",
                     "Prosocial agents",
                     #"Symbolic play", 
                     "Infant directed speech preference") ~ "Communication", 
      
      ds_name %in% c("Abstract rule learning",  
                     "Categorization bias", 
                     "Simple arithmetic competences") ~ "Cognitive"
     )
  ) %>% 
  mutate(pred = case_when(
    ds_name == "Abstract rule learning" ~ -pred, 
    ds_name == "Switch task" ~ -pred,
    TRUE ~ pred
  )) %>% 
  select(ds_name, ds_type, mean_age_months, pred, ci.ub, ci.lb) %>% 
  mutate(
    ds_name = case_when(
      ds_name %in% sig_ds ~ paste0(ds_name, "*"), 
      TRUE ~ ds_name
    ), 
    sig_alpha = case_when(
      ds_name %in% sig_ds  ~ 1, 
      TRUE ~ .3
    )
  )

categorized_age_pred_df %>% 
  # this is to handle the tricky part with neonatal imitation
  filter(pred != -Inf) %>% 
  mutate(group_var = paste0(ds_name, model_type)) %>% 
  mutate(
     ds_name = case_when(
      ds_name == "Cross-situational word learning" ~ "X-situation learning", 
      ds_name == "Label advantage in concept learning" ~ "Label advantage",
      ds_name == "Language discrimination and preference (discrimination)" ~ "Language discrimination", 
      ds_name == "Language discrimination and preference (preference)" ~ "Language preference",
      ds_name == "Cross-situational word learning"  ~ "X-situation learning",
      ds_name == "Gaze following (combined)" ~ "Gaze following", 
      ds_name == "Label advantage in concept learning" ~ "Label advantage", 
      ds_name == "Statistical sound category learning" ~ "Stats. category learning",
      ds_name == "Statistical word segmentation" ~ "Stats. word seg.",
      ds_name == "Word Segmentation (combined)" ~ "Word segmentation", 
      ds_name == "Infant directed speech preference" ~ "IDS pref.",
      TRUE ~ ds_name
  )) %>% 
  mutate(
    ds_name = case_when(
      ds_name %in% significant_res ~ paste0(ds_name, "*"), 
      TRUE ~ ds_name
    )
  ) %>% 
  ggplot(aes(x = mean_age_months,  y = pred)) + 
  geom_hline(yintercept = 0, linetype ="dotted", alpha = .5) + 
  geom_point(data = categorized_d , 
            aes(x = mean_age_months, y = d_calc),size = .5, alpha = .5) + 
  geom_line(alpha = 1,  aes(group = group_var,color = model_type)) + 
  
 geom_ribbon(
              aes(x = mean_age_months, group = group_var,y = pred, ymin = ci.lb, ymax = ci.ub, fill = model_type), alpha = 0.2, colour = NA)+ 

    theme_few()+
    

xlab("Mean age (months)") + 
  ylab("Predicted ES") + 
  
   theme(legend.position = "none", 
         strip.text = element_text(size = 7, margin = margin())) + 
  facet_wrap(~ds_name, scales = "free_x",
             labeller = label_wrap_gen()) 
```


```{r}
label_df %>% 
  mutate(
    x = case_when(
      ds_name == "Categorization bias" ~ x + 5, 
      TRUE ~ x
    ), 
    
    y = case_when(
      ds_name == "Categorization bias" ~ y + 1, 
      TRUE ~ y
    )
    
  )
```


@TODO: need fixing after redid the curves


```{r}



sig_ds <- c("Mutual exclusivity", "Audio-Visual Congruence", "Mispronunciation sensitivity",
           "Statistical sound category learning (habituation)", "Simple arithmetic competences")

categorized_age_pred_df <- all_age_pred %>% 
  select(pred, ds_name, ci.ub, ci.lb) %>% 
  group_by(ds_name) %>% 
  mutate(mean_age_months = (row_number() + 1) / 2) %>% 
  left_join(age_range_df, by = c("ds_name")) %>% 
  filter(mean_age_months > min_age, 
         mean_age_months < min(36, max_age)) %>% 
  mutate(
    ds_type = case_when(
      ds_name %in% c( "Language discrimination and preference", 
                     "Mispronunciation sensitivity", 
                     "Natural speech preference", "Sound symbolism", "Statistical sound category learning (habituation)", 
                     "Vowel discrimination (native)",
                     "Vowel discrimination (non-native)" ,"Word Segmentation (combined)",
                     "Statistical word segmentation", "Audio-Visual Congruence") ~ "Sounds", 
      
      ds_name %in% c("Familiar word recognition", 
                     "Cross-situational word learning", 
                      "Label advantage in concept learning",
                     "Mutual exclusivity",
                     "Online word recognition",
                     "Syntactic bootstrapping",
                     "Switch task") ~ "Words",
      
      ds_name %in% c("Gaze following (combined)",
                     "Neonatal Imitation",
                     "Prosocial agents",
                     #"Symbolic play", 
                     "Infant directed speech preference") ~ "Communication", 
      
      ds_name %in% c("Abstract rule learning",  
                     "Categorization bias", 
                     "Simple arithmetic competences") ~ "Cognitive"
     )
  ) %>% 
  mutate(pred = case_when(
    ds_name == "Abstract rule learning" ~ -pred, 
    ds_name == "Switch task" ~ -pred,
    TRUE ~ pred
  )) %>% 
  select(ds_name, ds_type, mean_age_months, pred, ci.ub, ci.lb) %>% 
  mutate(
    ds_name = case_when(
      ds_name %in% sig_ds ~ paste0(ds_name, "*"), 
      TRUE ~ ds_name
    ), 
    sig_alpha = case_when(
      ds_name %in% sig_ds  ~ 1, 
      TRUE ~ .3
    )
  )



label_df <- categorized_age_pred_df %>% 
  filter(mean_age_months == max(mean_age_months)) 

label_df$x = c(
    13 + 4,#Abstract rule learning
    0 + 8, # audio visual
    35.5 + 5 , # Categorization bias
    35.5 + 3, # "Cross-situational word learning"
    15 + 3,#Familiar word recognition
    23.5 + 4, #"Gaze following (combined)",,
    15 - 3, # Label advantage
    11.5 + 8, # "Language discrimination and preference"
    35.5 + 4,#"Mispronunciation sensitivity"
    35.5 + 4, #Mutual exclusivity , 
    12.5 , #"Natural speech preference"
    0 + 4, #neonatal imitation 
    30 + 2, #"Online word recognition"
    31.5 + 8, #Prosocial Agents
    9 + 4, #"Simple arithmetic competences"
    35.5 + 8, #Sound symbolism, 
    11 + 5, #"Statistical sound category learning (habituation)"
    18.5, #"Statistical word segmentation"
    24.5+ 3, #Switch task,
    35.5 + 5, # "Syntactic bootstrapping"
    17.5 - 5, #"Vowel discrimination (non-native)"
    29.5 + 11, #"Vowel discrimination (native)"
    24.5 + 14, #"Word Segmentation (combined)"
    18.5# "Infant directed speech preference"
)
label_df$y = c(
  0.05 - 0.5,#Abstract rule learning
  0 + 1.5, # audio visual
  0.275 + .4, # Categorization bias
  0.545 + .3, # "Cross-situational word learning"
  0.805 + 0.3,#Familiar word recognition
  2.99 + 0.3, #"Gaze following (combined)"
  0.05004268 - 0.5, # label advantage
  0.265 - 1, # "Language discrimination and preference"
  1.62 + 0.5,#"Mispronunciation sensitivity"
  1.44 + 0.3,#Mutual exclusivity
  0.811 + 1.3, #"Natural speech preference"
  0 + 1.5, # neonatal imitation 
  2.964331 + 0.5, #"Online word recognition"
  0.405, #Prosocial agents
  0.249 + 0.6,#"Simple arithmetic competences"
  0.383, #Sound symbolism, 
  0.557 + 0.5, #"Statistical sound category learning (habituation)"
  -0.07808538 - 0.2, #"Statistical word segmentation"
  0.156 - 0.3,#Switch task
  0.103 - 0.5, # "Syntactic bootstrapping"
  0.5901183 + 0.2, #"Vowel discrimination (native)"
  0.6457766 + 1, #"Vowel discrimination (non-native)"
  0.2 + 1.2, #"Word Segmentation (combined)"
 0.467 - 0.6# "Infant directed speech preference", 
 #0.627 + 0.5, #Symbolic play
)

p <- categorized_age_pred_df %>% 
  ggplot(aes(x = mean_age_months, y = pred, group = ds_name, color = ds_name)) + 
  geom_hline(yintercept = 0, linetype = "dashed", alpha = .5) + 
  geom_line(aes(alpha = sig_alpha), width = 2)+ 
  geom_ribbon(
              aes(x = mean_age_months, y = pred, ymin = ci.lb, ymax = ci.ub, fill = ds_name), alpha = 0.1, colour = NA)+ 
  geom_text(data = label_df, aes(x = x, y = y, label = ds_name), 
                   size = 2.2,
                  segment.color = NA) + 
  xlim(0, 48) + 
  ylim(-2, 4) + 
  facet_wrap(~ds_type, scales = "free") + 
  theme_few()+
  theme(legend.position = "none") + 
  xlab("Age (months)") + 
  ylab("Predicted Value")
  

p
```






# Misc

## cf const 

```{r}
age_df_wide %>% 
  filter(Const == 0) %>% 
  pivot_longer(cols = c("Linear", "Log", "Quadratic", "Const")) %>% 
  ggplot(aes(x = reorder(Dataset, value), y = value, color = name)) +
  theme_few() + 
  geom_hline(yintercept = 4, linetype = 2, color = "gray") + 
  geom_hline(yintercept = 0, linetype = 3) + 
  geom_point(position = position_dodge(width = .2), 
             #color = "white"
             ) + 

  coord_flip()+ 
  xlab("") + 
  labs(color = "Functional form") + 
  guides(colour = guide_legend(nrow = 2))  + 
  theme(legend.position = "top")
```

linear_model_formula <-"d_calc ~ mean_age_months"
log_model_formula <- "d_calc ~ log(mean_age_months)"
qua_model_formula <- "d_calc ~ I(mean_age_months^2)"
const_model_formula <- "d_calc ~ 1"
    
    
```{r}
ds_name =

linear_model_formula <-"d_calc ~ mean_age_months"
log_model_formula <- "d_calc ~ log(mean_age_months)"
qua_model_formula <- "d_calc ~ I(mean_age_months^2)"
const_model_formula <- "d_calc ~ 1"

m1 <- rma.mv(as.formula(linear_model_formula), 
                                       V = d_var_calc, 
                                             random = ~ 1 | short_cite/unique_row, 
                                             data = d %>% filter(ds_clean == "Abstract rule learning"), 
             method="ML")

m2 <- rma.mv(as.formula(const_model_formula), 
                                       V = d_var_calc, 
                                             random = ~ 1 | short_cite/unique_row, 
                                             data = d %>% filter(ds_clean == "Abstract rule learning"), 
             method="ML")


anova(m1, m2)$pval
```

```{r}

get_model_type <- function(s){
  if (grepl("log", s)){
    return ("log")
  }else if (grepl("2", s)){
    return ("quad")
  }else{
    return ("linear")
  }
}

get_const_comparion <- function(all_ds, ds_name){
  
  const_m <- rma.mv(d_calc ~ 1, 
                  V = d_var_calc, 
                  random = ~ 1 | short_cite/same_infant/unique_row, 
                  data = all_ds %>% filter(ds_clean == ds_name), 
                  method = "ML")
  
  linear_model_formula <-"d_calc ~ mean_age_months"
  log_model_formula <- "d_calc ~ log(mean_age_months)"
  qua_model_formula <- "d_calc ~ I(mean_age_months^2)"
  
  res = lapply(c(linear_model_formula, 
                 log_model_formula, 
                 qua_model_formula), 
               function(f){
                 
                 m <- rma.mv(as.formula(f), 
                             V = d_var_calc, 
                             random = ~ 1 | short_cite/same_infant/unique_row, 
                              data = all_ds %>% filter(ds_clean == ds_name), 
                             method = "ML")
                 
                m_type <- get_model_type(f)
                
                res <- anova(m, const_m)
                return (tibble(
                  "m_type" = m_type, 
                  "pval" = res$pval
                ))
                 
                 
                 
                 
               }) %>% bind_rows()
  
  return(res %>% mutate(ds_name = ds_name))
  
  
}


compare_const_df <- lapply(unique(d$ds_clean), 
       function(name){
         print(name)
         get_const_comparion(d, name)
       }) %>% 
  bind_rows()

saveRDS(compare_const_df, here("cached_data/compare_const_df.Rds"))

```

    
```{r}
compare_const_df <- readRDS( here("cached_data/compare_const_df.Rds"))
compare_const_df %>% 
  ggplot(aes(x = ds_name, y = pval, color = m_type)) + 
  geom_point(position = position_dodge(width = .2), alpha = .3) + 
  theme_few() +
  geom_hline(yintercept = 0.05) +  
  coord_flip()
```

## Get predictions 

```{r}

get_all_age_model_prediction <- function(ds_name, min_age_df, formula, model_type, all_ds){
  
  current_df = all_ds %>% filter(ds_clean == ds_name)
  model <- rma.mv(as.formula(formula), 
                    V = d_var_calc, 
                    random = ~ 1 | short_cite/same_infant/unique_row, 
                    data = current_df)
    
 
  
  if (model_type == "Log"){
    age = log(seq(0, 36, .5))
  }else if(model_type == "Quadratic"){
    age = (seq(0, 36, .5)) ** 2
  }else{
    age = seq(0, 36, .5)
  }
  
  if (model_type != "Const"){
    prediction =  predict(model, newmods = age, addx =TRUE) %>% 
      as.data.frame() %>% 
      mutate(ds_name = ds_name)
    
  }else{
    
    prediction = predict(model) %>% 
      as.data.frame() 
    
    prediction = map_dfr(seq(length(age)), ~prediction)
    
    prediction$X.const.mean_age_months = age
    
    
    prediction$ds_name = ds_name
    
  }

  
  return (prediction)
}


get_age_model_prediction <- function(ds_name, min_age_df, all_ds){
  
  current_df = all_ds %>% filter(ds_clean == ds_name)
  formula = (min_age_df %>% filter(dataset == ds_name))$model_spec
  model_type = (min_age_df %>% filter(dataset == ds_name))$model_spec_clean
  model <- rma.mv(as.formula(formula), 
                    V = d_var_calc, 
                    random = ~ 1 | short_cite/same_infant/unique_row, 
                    data = current_df)
    
 
  
  if (model_type == "Log"){
    age = log(seq(0, 36, .5))
  }else if(model_type == "Quadratic"){
    age = (seq(0, 36, .5)) ** 2
  }else{
    age = seq(0, 36, .5)
  }
  
  if (model_type != "Const"){
    prediction =  predict(model, newmods = age, addx =TRUE) %>% 
      as.data.frame() %>% 
      mutate(ds_name = ds_name)
    
  }else{
    
    prediction = predict(model) %>% 
      as.data.frame() 
    
    prediction = map_dfr(seq(length(age)), ~prediction)
    
    prediction$X.const.mean_age_months = age
    
    
    prediction$ds_name = ds_name
    
  }

  
  return (prediction)
}

age_range_df <- d %>% 
  group_by(ds_clean) %>% 
  summarise(min_age = min(mean_age_months, na.rm = TRUE), 
            max_age = max(mean_age_months, na.rm = TRUE)) %>% 
  rename(ds_name = ds_clean)

ds_names = unique(d$ds_clean)
model_type = c("Log", "Quadratic", "Const", "Linear")
age_models_df <- readRDS(here("cached_data/age_models_df.RDS")) %>% 
  filter(ic == "AICc") %>% 
  group_by(dataset) %>% 
  filter(REML == min(REML))

pred_res <- lapply(ds_names, function(name){
  
  get_age_model_prediction(name, age_models_df, d)
}) %>% 
  bind_rows()
```


```{r}



ids_pred <- ids_all_model_df %>% 
  select(pred, ds_name, model_type, ci.lb, ci.ub) %>% 
  group_by(ds_name,model_type) %>% 
  mutate(mean_age_months = (row_number() + 1) / 2) %>% 
  left_join(age_range_df, by = c("ds_name")) %>% 
  filter(mean_age_months > min_age, 
         mean_age_months < min(36, max_age)) %>% 

  select(ds_name,  mean_age_months, pred, ci.lb, ci.ub)
```
