==> analysis/02_pubbias.rmd <==
---
title: "02_pubbias.rmd"
author: "Anjie Cao, Molly Lewis, and Michael Frank"
date: "`r Sys.Date()`"
output: html_document
---

```{r}
library(tidyverse)
library(here)
library(ggthemes)
library(metafor)
library(PublicationBias)
library(lme4)
library(lmerTest)
library(weightr)

source(here("analysis/helper/pubbias_correction.R"))


d <- read_csv(here("data/metalab_data_kitchensink.csv"))
```


```{r}

bias_lrt <- lapply(
  unique(d$ds_clean), 
  function(name){
    tryCatch(get_pub_bias_lrt(d, name), 
             warning = function(w){
               tibble("ds_clean" = name, 
                      "warning" = as.character(w))
             })
  }
) %>% bind_rows()

bias_estimates <- lapply(
  unique(d$ds_clean), 
  function(name){
    tryCatch(get_pub_bias_estimates(d, name), 
             warning = function(w){
               tibble("ds_clean" = name, 
                      "warning" = as.character(w))
             })
  }
) %>% bind_rows()

```
```{r}
bias_lrt_12mo <- lapply(
  unique(d$ds_clean), 
  function(name){
    print(name)
    tryCatch(get_pub_bias_lrt(d, name, split_option = "12mo"), 
             warning = function(w){
               tibble("ds_clean" = name, 
                      "warning" = as.character(w))
             })
  }
) %>% bind_rows()

bias_estimates_12mo <- lapply(
  unique(d$ds_clean), 
  function(name){
    print(name)
    tryCatch(get_pub_bias_estimates(d, name, split_option = "12mo"), 
             warning = function(w){
               tibble("ds_clean" = name, 
                      "warning" = as.character(w))
             })
  }
) %>% bind_rows()
```



```{r}
bias_lrt %>% 
```

```{r}
bias_estimates %>% 
  filter(test_data == "full") %>% 
  filter(moderator == "mean_age_months") %>% 
  ggplot(aes(x = type, y = estimate)) + 
  geom_point() + 
  geom_line(aes(group = ds_clean), alpha = .3)+
  stat_summary(col = "red") + 
  #facet_wrap(~ds_clean) + 
  #geom_hline(yintercept = 0.05, lty = 2) + 
  theme_few()
```


```{r}
bias_lrt %>% 
  filter(test_data != "full") %>% 
  ggplot(aes(x = test_data, y = pvalue)) + 
  geom_point() + 
  geom_line(aes(group = ds_clean), alpha = .3)+
  stat_summary(col = "red") + 
  #facet_wrap(~ds_clean) + 
  geom_hline(yintercept = 0.05, lty = 2) + 
  theme_few()
```


```{r}
d_test <- d %>% filter(ds_clean == "Mutual exclusivity")

res <- rma.mv(d_calc ~ 1,
              V = d_var_calc, 
               random = ~ 1 | short_cite/unique_row, 
                data = d_test)


```


```{r}
# this function tests whether there is publication bias present 

get_lrt <- function(model){
  df <- length(model[[2]]$par) - length(model[[1]]$par)
  lrchisq <- 2*(abs(model[[1]]$value - model[[2]]$value))
  pvalue <- 1-pchisq(lrchisq,df)
  
  return (
    tibble(
      "lrchisq" = lrchisq, 
      "df" = df, 
      "pvalue" = pvalue
    )
  )
}


get_estimates <- function(model){
  
  adjusted_estimates = tibble(
    "moderator" = c("intercept", "mean_age_months"),
    "estimate" = c(model$adj_est[[2]], model$adj_est[[3]]), 
    "ci.lb" = c(model$ci.lb_adj[[2]], model$ci.lb_adj[[3]]), 
    "ci.ub" = c(model$ci.ub_adj[[2]], model$ci.ub_adj[[3]]),
    "se" = c(model$adj_se[[2]], model$adj_se[[3]]), 
    "z" = c(model$z_adj[[2]], model$z_adj[[3]]), 
    "p_val" = c(model$p_adj[[2]], model$p_adj[[3]])
  ) %>% mutate(type = "adjusted")
  
  
  unadjusted_estimates = tibble(
    "moderator" = c("intercept", "mean_age_months"),
    "estimate" = c(model$unadj_est[[2]], model$unadj_est[[3]]), 
    "ci.lb" = c(model$ci.lb_unadj[[2]], model$ci.lb_unadj[[3]]), 
    "ci.ub" = c(model$ci.ub_unadj[[2]], model$ci.ub_unadj[[3]]),
    "se" = c(model$unadj_se[[2]], model$unadj_se[[3]]), 
    "z" = c(model$z_unadj[[2]], model$z_unadj[[3]]), 
    "p_val" = c(model$p_unadj[[2]], model$p_unadj[[3]])
  ) %>% mutate(type = "unadjusted")
  
  return (bind_rows(
    adjusted_estimates, unadjusted_estimates
  ))
  
  
  
}

# separate datasets into two


mean_age <- mean(d_test$mean_age_months)
younger_ds <- d_test %>% filter(mean_age_months >= mean_age)
older_ds <- d_test %>% filter(mean_age_months < mean_age)

# check how many studies are in each half 
n_younger <- nrow(younger_ds)
n_older <- nrow(older_ds)

# run three models 
# first half 
younger_res <- weightr::weightfunct(effect = younger_ds$d_calc, 
                     v = younger_ds$d_var_calc, 
                     steps = c(0.05, 1), 
                     mods = ~ younger_ds$mean_age_months 
                     )

younger_p_val <- get_lrt(younger_res)
younger_estimates <- get_estimates(younger_res)

# second half 
older_res <- weightr::weightfunct(effect = older_ds$d_calc, 
                     v = older_ds$d_var_calc, 
                     steps = c(0.05, 1), 
                     mods = ~ older_ds$mean_age_months 
                     )
older_p_val <- get_lrt(older_res)
older_estimates <- get_estimates(older_res)

  
# together 
full_res <- weightr::weightfunct(effect = d_test$d_calc, 
                     v = d_test$d_var_calc, 
                     steps = c(0.05, 0.5, 1), 
                     mods = ~ d_test$mean_age_months 
                     )
full_p_val <- get_lrt(full_res)
full_estimates <- get_estimates(full_res)


lrt_summary <- bind_rows(
  younger_p_val %>% mutate(test_data = "younger"), 
  older_p_val %>% mutate(test_data = "older"), 
  full_p_val %>% mutate(test_data = "full")
)

estimates_summary <- bind_rows(
  younger_estimates %>% mutate(test_data = "younger"), 
  older_estimates %>% mutate(test_data = "older"), 
  full_estimates %>% mutate(test_data = "full")
)

                     

res = 
          

```



# Egger's test 

## Caching results 


```{r}

get_tidy_regtest <- function(regtest_res){
  tibble(
    "zval" = regtest_res$z_val, 
    "pval" = regtest_res$pval,
    "limit_est" = regtest_res$est, 
    "lb" = regtest_res$ci.lb,
    "ub" = regtest_res$ci.ub
  )
}

age_summary <- d %>% 
  group_by(ds_clean) %>% 
  summarise(mid_point = median(mean_age_months))

egger_test_ok <- d %>% 
  left_join(age_summary, by = c("ds_clean")) %>% 
  mutate(
    age_group = case_when(
      mean_age_months < mid_point ~ "younger", 
      TRUE ~ "older"
    )
  ) %>%
  group_by(ds_clean, age_group) %>% 
  count() %>% 
  filter(n >= 20) %>% 
  group_by(ds_clean) %>% 
  count() %>% 
  filter(n == 2)


egger_test_df <- d %>% 
  left_join(age_summary, by = c("ds_clean")) %>% 
  mutate(
    age_group = case_when(
      mean_age_months < mid_point ~ "younger", 
      TRUE ~ "older"
    )
  ) %>%
  group_by(ds_clean, age_group) %>% 
  filter(ds_clean %in% egger_test_ok$ds_clean) %>% 
  nest() %>% 
  mutate(
    reg_fit = map(data, ~ get_tidy_regtest(regtest(x = .x$d_calc, vi = .x$d_var_calc)))
  ) %>% 
  unnest(reg_fit)

saveRDS(egger_test_df, here("cached_data/h1_egger_test.rds"))

egger_test_df <- readRDS(here("cached_data/h1_egger_test.rds"))



```


## Visualization

```{r}
egger_test_df
```

# Egger vs Bias 

```{r}
egger_test_df %>%
  select(ds_clean, age_group, limit_est, pval) %>% 
  pivot_wider(names_from = age_group, 
                              values_from = c(limit_est, pval)) %>% 
  mutate(
    younger_sig = pval_younger < .05, 
    older_sig = pval_older < .05
  ) %>% 
  mutate(
    comparison_direction = case_when(
      (younger_sig == TRUE & older_sig == FALSE) ~ "younger_more_severe", 
      (younger_sig == FALSE & older_sig == TRUE) ~ "older_more_severe", 
      (younger_sig == TRUE & older_sig == TRUE) ~ "both_pub_bias", 
      (younger_sig == FALSE & older_sig == FALSE) ~ "no pub bias"
    )
  ) %>% 
  saveRDS(here("cached_data/pub_bias_egger_test.Rds"))

#%>% 
#  filter(comparison_direction == "younger_more_severe")

egger_sig <- egger_test_df %>%
  select(ds_clean, age_group, limit_est, pval) %>% 
  pivot_wider(names_from = age_group, 
                              values_from = c(limit_est, pval)) %>% 
  mutate(
    younger_sig = pval_younger < .05, 
    older_sig = pval_older < .05
  ) %>% 
  mutate(
    comparison_direction = case_when(
      (younger_sig == TRUE & older_sig == FALSE) ~ "younger_more_severe", 
      (younger_sig == FALSE & older_sig == TRUE) ~ "older_more_severe", 
      (younger_sig == TRUE & older_sig == TRUE) ~ "both_pub_bias", 
      (younger_sig == FALSE & older_sig == FALSE) ~ "no pub bias"
    )
  ) %>% 
  filter(comparison_direction == "younger_more_severe")

egger_test_df %>%
  select(ds_clean, age_group, limit_est, pval) %>% 
  pivot_wider(names_from = age_group, 
                              values_from = c(limit_est, pval)) %>% 
  mutate(
    younger_sig = pval_younger < .05, 
    older_sig = pval_older < .05
  ) %>% 
  mutate(
    comparison_direction = case_when(
      (younger_sig == TRUE & older_sig == FALSE) ~ "younger_more_severe", 
      (younger_sig == FALSE & older_sig == TRUE) ~ "older_more_severe", 
      (younger_sig == TRUE & older_sig == TRUE) ~ "both_pub_bias", 
      (younger_sig == FALSE & older_sig == FALSE) ~ "no pub bias"
    )
  ) %>% 
  #filter(comparison_direction == "younger_more_severe") %>% 
  filter(ds_clean %in% c("Mutual exclusivity", "Vowel discrimination (non-native)"))
```



```{r}
bias_lrt %>% 
  filter(is.na(warning)) %>% 
  select(ds_clean, test_data,  pvalue, lrchisq) %>% 
  pivot_wider(names_from = test_data, 
                              values_from = c(pvalue, lrchisq)) %>% 
  mutate(
    younger_sig = pvalue_younger < .05, 
    older_sig = pvalue_older < .05
  ) %>% 
  mutate(
    comparison_direction = case_when(
      (younger_sig == TRUE & older_sig == FALSE) ~ "younger_more_severe", 
      (younger_sig == FALSE & older_sig == TRUE) ~ "older_more_severe", 
      (younger_sig == TRUE & older_sig == TRUE) ~ "both_pub_bias", 
      (younger_sig == FALSE & older_sig == FALSE) ~ "no pub bias"
    )
  ) %>% 
  saveRDS(here("cached_data/pub_bias_weight_function.Rds"))

#%>% 
  #filter(ds_clean %in% egger_sig$ds_clean)
  #filter(comparison_direction == "younger_more_severe")
  #filter(ds_clean %in% c("Audio-Visual Congruence", "Categorization bias", "Syntactic bootstrapping"))
```

```{r}
egger_test_df %>% 
  ggplot(aes(x = age_group, y = pval)) + 
  geom_point() + 
  geom_line(aes(group = ds_clean), alpha = .3)+
  stat_summary(col = "red") + 
  #facet_wrap(~ds_clean) + 
  geom_hline(yintercept = 0.05, lty = 2) + 
  theme_few() + 
  xlab("Age Group") + 
  ylab("p value")


```


```{r}
readRDS(here("cached_data/pub_bias_weight_function.Rds"))
```

# Try 12 months old split 


```{r}

```


# Attrition? 

## Running stats 

```{r}
#~ 1 | short_cite/same_infant/unique_row

exclusion_d <- d %>% 
  filter(!is.na(n_excluded_1)) %>% 
  filter(!is.na(n_excluded_1)) %>% 
  mutate(
    p_exclusion = n_excluded_1 / (n_excluded_1 + n_1)
  ) %>% 
  select(ds_clean, short_cite, mean_age_months, p_exclusion) %>% 
  group_by(ds_clean) %>% 
  nest()

exclusion_d %>% 
  mutate(model = map(data, ~broom.mixed::tidy(lmer(p_exclusion ~ mean_age_months + (1 | short_cite), data = .), effects = "fixed", conf.int=TRUE, conf.level = 0.95))) %>% 
  unnest(model) %>% 
  filter(term == "mean_age_months") %>% 
  write_rds(here("cached_data/attrition_data.Rds"))
  
```



## Visualization 

```{r}
exclusion_d <- readRDS(here("cached_data/attrition_data.Rds"))


exclusion_d %>% 
  filter(conf.low != -Inf) %>% 
  mutate(
    sig = if_else(p.value < .05, TRUE, FALSE)
  ) %>% 
  ggplot(aes(x = ds_clean, y = estimate, ymin = conf.low, ymax = conf.high)) + 
  geom_pointrange(aes(color = sig)) + 
  geom_hline(yintercept = 0) + 
  coord_flip() + 
  theme_few()
```




```{r}
d %>% 
  filter(!is.na(n_excluded_1)) %>% 
  mutate(
    p_exclusion = n_excluded_1 / (n_excluded_1 + n_1)
  ) %>% 
  ggplot(aes(x = mean_age_months, y = p_exclusion)) + 
  geom_point() + 
  geom_smooth(method = "lm") + 
  facet_wrap(~ds_clean, scales = "free") 
```









# Misc


```{r}

pubbias_estimate <- function(d, ratio){
  
  print(d$long_cite[[1]])
  print(ratio)
  if(!is.na(d$long_cite[[1]])){
      cluster_var = d$long_cite
  }else{
    cluster_var = d$study_ID
  }
  
  if (cluster_var[[1]] == "Bahrick, L. E., & Pickens, J. N. (1988). Classification of bimodal English and Spanish language passages by infants. Infant Behavior and Development, 11(3), 277-296. https://doi.org/10.1016/0163-6383(88)90014-8" | cluster_var[[1]] ==  "Altvater-Mackensen, N., & Fikkert, P. (2010). The acquisition of the stop-fricative contrast in"){
    favor_dir = FALSE
  }else{
    favor_dir = TRUE
  }
  
  res <- mod_pubbias_meta(
  yi = d$d_calc, 
  vi = d$d_var_calc, 
  cluster = cluster_var,
  selection_ratio = ratio, 
  model_type = "robust", 
  favor_positive = favor_dir
  )
  
  res <- res$stats %>% mutate(ratio = ratio)
  
  return(res)
  
}


```


```{r}

nested_d <- d %>% 
  group_by(ds_clean) %>% 
  nest(data = -ds_clean) %>% 
  crossing(
    tibble(ratio = seq(1, 4, 0.2))
  )


sim_pub_ratio_df <- nested_d %>% 
  mutate(res = map2_df(data, ratio, pubbias_estimate))

```


```{r}
saveRDS(sim_pub_ratio_df, here("cached_data/pubbias_sim.Rds"))
```



```{r}
tidy_sim_pub_ratio_df <- sim_pub_ratio_df %>% select(-c(ratio, data)) %>% unnest(res) %>%
  mutate(sig_p = (p_value < 0.05))
  
```

```{r}
tidy_sim_pub_ratio_df %>% 
ggplot(aes(x = ratio, y = estimate)) + 
  geom_pointrange(aes(ymin = ci_lower, ymax = ci_upper, color = sig_p)) +
  geom_hline(yintercept = 0) + 
  theme_few() + 
  facet_wrap(~ds_clean)

```
