==> analysis/05_infant_constant.Rmd <==
```{r}
library(tidyverse)
library(here)
library(metafor)

source(here("analysis/helper/model_comparison_help.r"))

d <- read_csv(here("data/metalab_data_mini.csv"))
full_comparison <- readRDS(here("cached_data/age_models_df.Rds"))
```

# Caching results 

```{r}
# first take out mostly younger infants

too_young_domain <- d %>% 
  group_by(ds_clean) %>% 
  summarise(max_age = max(mean_age_months)) %>% 
  filter(max_age < 12)

# looking at only 12 month and above
toddler_d_with_enough_data <- d %>% 
  filter(!ds_clean %in% too_young_domain$ds_clean) %>% 
  filter(mean_age_months > 12) %>% 
  group_by(ds_clean) %>% 
  count() %>% 
  filter(n > 10)

toddler_d <- d %>% 
  filter(ds_clean %in% toddler_d_with_enough_data$ds_clean) %>% 
  filter(mean_age_months > 12) 
  
full_d <- d %>% 
  filter(!ds_clean %in% too_young_domain$ds_clean) 

toddler_comparison <- lapply(
  unique(toddler_d$ds_clean), 
  function(ds_name){
    print(ds_name)
    get_compare_IC_df(toddler_d, ds_name, NULL)
  }
) %>% 
  bind_rows() %>% 
  mutate(ds_type = "toddler")



write_rds(toddler_comparison, here("cached_data/above_12_month_models_df.Rds"))
```


```{r}
all_toddler_slope_estimates <- lapply(unique(toddler_d$ds_clean), 
       function(name){
         print(name)
         get_model_fit_df(toddler_d, name, NULL)
       }) %>% 
  bind_rows() %>% 
  mutate(data_type = "all_toddler")

all_toddler_slope_estimates %>% write_rds(here('cached_data/all_toddler_slope_estimates.Rds'))

all_slope_estimates <- read_rds(here("cached_data/all_slope_estimates.Rds")) %>% 
  mutate(data_type = "full") %>% 
  filter(dataset %in% all_toddler_slope_estimates$dataset)

all_toddler_slope_estimates %>% 
  bind_rows(all_slope_estimates) %>% 
  filter(term != "intercept") %>% 
  ggplot(aes(x = reorder(dataset, estimate), y = estimate, ymin = conf.low, ymax = conf.high, color = data_type, group = data_type)) + 
  scale_color_manual(values = c("black", "red")) + 
  geom_hline(yintercept = 0, linetype = "dashed") + 
  geom_pointrange(position = position_dodge(width = .8), alpha = .3) + 
  geom_line(aes(group = dataset), po)+ 
  coord_flip() + 
  theme_few() + 
  xlab("")


all_toddler_slope_estimates %>% 
  bind_rows(all_slope_estimates) %>% 
  mutate(is_sig = p.value < 0.05) %>% 
  filter(term != "intercept") %>% 
  ggplot(aes(x = data_type, y = estimate, ymin = conf.low, ymax = conf.high, color = is_sig)) + 
  geom_pointrange() +
  facet_wrap(~dataset, scales = "free")
```




# Visualization 


```{r}
all_toddler_slope_estimates <- read_rds(here('cached_data/all_toddler_slope_estimates.Rds')) %>% 
  mutate(data_type = "12MO+ only") %>% 
  filter(term == "mean_age_months") %>% 
  filter(dataset != "Audio-Visual Congruence") 
all_slope_estimates <- read_rds(here("cached_data/all_slope_estimates.Rds")) %>% 
  mutate(data_type = "Full dataset") %>% 
  filter(term == "mean_age_months")

bind_rows(all_toddler_slope_estimates, all_slope_estimates) %>% 
  ggplot(aes(x = dataset, y = estimate, ymin = conf.low, ymax = conf.high, color = data_type)) +
  geom_pointrange(position = position_dodge(width = .5)) + 
   geom_hline(yintercept = 0, linetype = 2, color = "gray") + 
  coord_flip() + 
  theme_few() + 
  xlab("") + 
  ylab("Estimate on mean age") +
  scale_color_manual(values = c("purple", "lightblue"), name = "") + 
  theme(legend.position = "top")
```


```{r}
wide_toddler <- toddler_comparison %>% 
  filter(ic == "AICc") %>% 
  filter(ic == "AICc") %>% 
  rename(Dataset = dataset) %>% 
  select(REML, model_spec_clean, Dataset) %>% 
  pivot_wider(names_from = model_spec_clean, 
              values_from = REML)

wide_toddler$min <- apply(wide_toddler[c('Linear','Log', 'Quadratic', 'Const')], 1, min)

# get the delta value 
wide_toddler <- wide_toddler %>% 
  mutate(d_linear = Linear - min, 
         d_log = Log - min, 
         d_quad = Quadratic - min, 
         d_const= Const - min) %>% 
  select(Dataset, d_linear, d_log, d_quad, d_const) %>% 
  rename(Linear = d_linear, 
         Log = d_log, 
         Quadratic = d_quad, 
         Const = d_const) %>% 
   mutate(across(where(is.numeric), round, 2)) %>% 
  mutate(ds_type = "12mon+")


wide_full <- full_comparison %>% 
  filter(ic == "AICc") %>% 
  filter(ic == "AICc") %>% 
  rename(Dataset = dataset) %>% 
  select(REML, model_spec_clean, Dataset) %>% 
  pivot_wider(names_from = model_spec_clean, 
              values_from = REML)

wide_full$min <- apply(wide_full[c('Linear','Log', 'Quadratic', 'Const')], 1, min)

# get the delta value 
wide_full <- wide_full %>% 
  mutate(d_linear = Linear - min, 
         d_log = Log - min, 
         d_quad = Quadratic - min, 
         d_const= Const - min) %>% 
  select(Dataset, d_linear, d_log, d_quad, d_const) %>% 
  rename(Linear = d_linear, 
         Log = d_log, 
         Quadratic = d_quad, 
         Const = d_const) %>% 
   mutate(across(where(is.numeric), round, 2)) %>% 
  mutate(ds_type = "full")


bind_rows(wide_toddler, wide_full) %>% 
  pivot_longer(cols = c("Linear", "Log", "Quadratic", "Const"), 
               names_to = "model_type", 
               values_to = "AICc") %>% 
  ggplot(aes(x = Dataset, y = AICc, color = model_type, shape = ds_type, group = ds_type)) + 
  geom_point(position = position_dodge(width = .8))+ 
  geom_hline(yintercept = 4, linetype = "dashed", alpha = .4, color = "gray")+
  theme_few() + 
  coord_flip()
```


```{r}
bind_rows(wide_toddler, wide_full) %>% 
  pivot_longer(cols = c("Linear", "Log", "Quadratic", "Const"), 
               names_to = "model_type", 
               values_to = "AICc") %>% 
  ggplot(aes(x = ds_type, y = AICc)) + 
  geom_point()+
  geom_line(aes(group = Dataset), alpha = .3) + 
  geom_hline(yintercept = 4, linetype = "dashed", alpha = .4, color = "gray")+
  theme_few() +
  stat_summary(col = "red") + 
  facet_wrap(~model_type) +
  ylab("Delta AIC")
```
